comment: add "Script Type:".
The "script type" is just the directory name of where the script resides
in.  So for a script in "foo/xxx.vim", its script type is then assigned
to be "foo".

I hope this can be useful in the future, but I don't exactly know for
sure *how* this can be useful...  Mmm.