syntax: add initial regular expression matching
I have chosen to match against "./<insert_your_regexp_here>/" although
dot-slash is technically just an ordinary Mythryl String, since the
dot-slash is the closest to the canonical "/<insert_regexp_here>/", and
is also used in the tutorials to represent a "regex".

Note that a lot of work will have to be done on matching against regexp
escapes, grouping, etc.