indent & snippets: add support for (de-)indent for braces and equal sign
Finally!

In order to support de-indentation for both record-braces and
codeblock-braces, the patch is just using the same indent level for the
closing brace "}" as the line that contains the matching starting brace
"{".

It also allows arbitrary indentation level for the line after the start
brace "{", i.e. not just the 2 & 4 blanks (respectively, for records and
code blocks).

In addition, support for the following style of indent with equal sign
is implemented:

	my foo
	    =
	    { ... }
	    bar baz;

	# de-indent here!

	fun foo'
	    =
	    if ...
	    elif ...
	    fi;

	# de-indent here!
	...

Furthermore, this patch turns the "GetMythrylIndent" function basically
into a huge if-elseif statement, which should get rid of a lot of
redundant condition evaluations and also ease debugging of trigger
conditions.

This in turn allows reordering of the "elseif" branches with regard only
to the trigger conditions and not the side-effects that the
elseif-statement blocks create, hence can be done with much less
frustration.