indent: hoist inner-ifs to outer-ifs; move comment tests to top
Hoisting the inner if-tests to the outer tests by "&&"-ing them makes
debugging test conditions easier.

Tests for comments are moved to the top and matches make the function
return as soon as possible.  The reason behind the move is that other
if-tests will skip over comments, which would have rendered the
comment-tests useless.

Also changed style of invocation of the function "cursor()" in the
custom function "PreviousMatching()"...  I have to admit I don't grok
the semantics of the "cursor()" function.  It seems to be doing funny
things sometimes, and changing the style of invocation seems to be able
to fix it.  I may be wrong, and may just be seeing side-effects from
other parts of the indent function.